refactor(media-modal): extract flow state/apply logic and update tests
